algorithm/sepsis_causal_project/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sepsis-causal"
version = "0.1.0"
description = "End-to-end causal transformer pipeline for sepsis multi-treatment simulation and evaluation."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "optuna>=4.0",
  "pandas>=2.0",
  "PyYAML>=6.0",
  "scikit-learn>=1.3",
  "torch>=2.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
